# contracts/openapi-v2.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PCConnect API",
    "version": "2.0.0",
    "description": "Authoritative REST contract for PCConnect controllers and agents. SignalR notifications are advisory; REST resources are authoritative."
  },
  "servers": [
    { "url": "https://api.pcconnect.adamdeveloping.co.uk/api/v2", "description": "Production" }
  ],
...
  "paths": {
...
    "/auth/password/change": {
      "post": {
        "tags": ["Identity"], "operationId": "changePassword", "parameters": [{ "$ref": "#/components/parameters/StepUpGrant" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PasswordChangeRequest" } } } },
        "responses": { "204": { "description": "Password changed and other sessions revoked" }, "400": { "$ref": "#/components/responses/Problem" }, "401": { "$ref": "#/components/responses/Problem" }, "403": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/auth/password/forgot": {
      "post": {
        "tags": ["Identity"], "operationId": "requestPasswordReset", "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EmailRequest" } } } },
        "responses": { "202": { "description": "Request accepted whether or not the account exists" }, "429": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/auth/password/reset": {
      "post": {
        "tags": ["Identity"], "operationId": "resetPassword", "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/PasswordResetRequest" } } } },
        "responses": { "204": { "description": "Password changed and all sessions revoked" }, "400": { "$ref": "#/components/responses/Problem" }, "410": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/auth/email/verify": {
      "post": {
        "tags": ["Identity"], "operationId": "verifyEmail", "security": [],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokenRequest" } } } },
        "responses": { "204": { "description": "Email verified" }, "400": { "$ref": "#/components/responses/Problem" }, "410": { "$ref": "#/components/responses/Problem" } }
      }
    },
...
    "/auth/step-up/options": {
      "post": {
        "tags": ["Identity"], "operationId": "createStepUpOptions",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StepUpIntent" } } } },
        "responses": { "200": { "description": "Allowed authentication methods and WebAuthn options", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StepUpOptions" } } } }, "400": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/auth/step-up/complete": {
      "post": {
        "tags": ["Identity"], "operationId": "completeStepUp",
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StepUpCompletion" } } } },
        "responses": { "200": { "description": "Intent-bound step-up grant", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StepUpGrant" } } } }, "400": { "$ref": "#/components/responses/Problem" }, "401": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/me": {
      "get": { "tags": ["Account"], "operationId": "getProfile", "responses": { "200": { "description": "Current profile", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Profile" } } } }, "401": { "$ref": "#/components/responses/Problem" } } },
      "patch": {
        "tags": ["Account"], "operationId": "updateProfile", "parameters": [{ "$ref": "#/components/parameters/StepUpGrant" }],
        "requestBody": { "required": true, "content": { "application/merge-patch+json": { "schema": { "$ref": "#/components/schemas/ProfileUpdate" } } } },
        "responses": { "200": { "description": "Updated profile", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Profile" } } } }, "400": { "$ref": "#/components/responses/Problem" }, "409": { "$ref": "#/components/responses/Problem" } }
      },
      "delete": {
        "tags": ["Account"], "operationId": "deleteAccount", "parameters": [{ "$ref": "#/components/parameters/StepUpGrant" }],
        "responses": { "202": { "description": "Deletion accepted; sessions and devices revoked" }, "401": { "$ref": "#/components/responses/Problem" }, "403": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/me/export": {
      "post": {
        "tags": ["Account"], "operationId": "requestDataExport", "parameters": [{ "$ref": "#/components/parameters/StepUpGrant" }],
        "responses": { "202": { "description": "Export job created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ExportJob" } } } }, "403": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/me/export/{exportId}": {
      "get": {
        "tags": ["Account"], "operationId": "getDataExport", "parameters": [{ "$ref": "#/components/parameters/ExportId" }],
        "responses": { "200": { "description": "Export status and short-lived URL when ready", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ExportJob" } } } }, "404": { "$ref": "#/components/responses/Problem" }, "410": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/me/export/{exportId}/download": {
      "get": {
        "tags": ["Account"], "operationId": "downloadDataExport", "parameters": [{ "$ref": "#/components/parameters/ExportId" }],
        "responses": { "200": { "description": "Decrypted short-lived user export", "content": { "application/json": { "schema": { "type": "object", "additionalProperties": true } } } }, "404": { "$ref": "#/components/responses/Problem" }, "410": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/sessions": {
      "get": {
        "tags": ["Account"], "operationId": "listSessions", "parameters": [{ "$ref": "#/components/parameters/Cursor" }, { "$ref": "#/components/parameters/Limit" }],
        "responses": { "200": { "description": "Sessions", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SessionPage" } } } } }
      }
    },
    "/sessions/{sessionId}": {
      "delete": {
        "tags": ["Account"], "operationId": "revokeSession", "parameters": [{ "$ref": "#/components/parameters/SessionId" }],
        "responses": { "204": { "description": "Session revoked" }, "404": { "$ref": "#/components/responses/Problem" } }
      }
    },
...
    "/devices/{deviceId}/commands": {
      "post": {
        "tags": ["Commands"], "operationId": "createCommand", "parameters": [{ "$ref": "#/components/parameters/DeviceId" }, { "$ref": "#/components/parameters/IdempotencyKey" }, { "$ref": "#/components/parameters/StepUpGrantOptional" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommandCreate" } } } },
        "responses": { "202": { "description": "Durably queued command", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Command" } } } }, "400": { "$ref": "#/components/responses/Problem" }, "403": { "$ref": "#/components/responses/Problem" }, "409": { "$ref": "#/components/responses/Problem" }, "422": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/commands": {
      "get": {
        "tags": ["Commands"], "operationId": "listCommands", "parameters": [{ "$ref": "#/components/parameters/Cursor" }, { "$ref": "#/components/parameters/Limit" }],
        "responses": { "200": { "description": "Owned command projections ordered by latest change", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommandPage" } } } } }
      }
    },
    "/commands/{commandId}": {
      "get": { "tags": ["Commands"], "operationId": "getCommand", "parameters": [{ "$ref": "#/components/parameters/CommandId" }], "responses": { "200": { "description": "Command status", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Command" } } } }, "404": { "$ref": "#/components/responses/Problem" } } }
    },
    "/commands/{commandId}/cancel": {
      "post": { "tags": ["Commands"], "operationId": "cancelCommand", "parameters": [{ "$ref": "#/components/parameters/CommandId" }], "responses": { "200": { "description": "Cancelled command", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Command" } } } }, "404": { "$ref": "#/components/responses/Problem" }, "409": { "$ref": "#/components/responses/Problem" } } }
    },
...
    "/agent/commands": {
      "get": {
        "tags": ["Agent"], "operationId": "listPendingCommands", "parameters": [{ "$ref": "#/components/parameters/Cursor" }, { "$ref": "#/components/parameters/Limit" }],
        "responses": { "200": { "description": "Unexpired commands for the authenticated device", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommandPage" } } } } }
      }
    },
    "/agent/commands/{commandId}/claim": {
      "post": {
        "tags": ["Agent"], "operationId": "claimCommand", "parameters": [{ "$ref": "#/components/parameters/CommandId" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommandClaim" } } } },
        "responses": { "200": { "description": "Claimed command with 30-second lease", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Command" } } } }, "404": { "$ref": "#/components/responses/Problem" }, "409": { "$ref": "#/components/responses/Problem" }, "410": { "$ref": "#/components/responses/Problem" } }
      }
    },
    "/agent/commands/{commandId}/acknowledgements": {
      "post": {
        "tags": ["Agent"], "operationId": "acknowledgeCommand", "parameters": [{ "$ref": "#/components/parameters/CommandId" }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CommandAcknowledgement" } } } },
        "responses": { "200": { "description": "Updated command", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Command" } } } }, "400": { "$ref": "#/components/responses/Problem" }, "409": { "$ref": "#/components/responses/Problem" } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer", "bearerFormat": "opaque" }
    },
    "parameters": {
      "Cursor": { "name": "cursor", "in": "query", "required": false, "schema": { "type": "string", "maxLength": 512 } },
      "Limit": { "name": "limit", "in": "query", "required": false, "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 50 } },
      "DeviceId": { "name": "deviceId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } },
...
      "SessionId": { "name": "sessionId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } },
      "CommandId": { "name": "commandId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } },
...
      "ExportId": { "name": "exportId", "in": "path", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } },
...
      "IdempotencyKey": { "name": "Idempotency-Key", "in": "header", "required": true, "schema": { "$ref": "#/components/schemas/Uuid" } },
      "StepUpGrant": { "name": "X-Step-Up-Grant", "in": "header", "required": true, "schema": { "type": "string", "minLength": 32, "maxLength": 512 } },
      "StepUpGrantOptional": { "name": "X-Step-Up-Grant", "in": "header", "required": false, "schema": { "type": "string", "minLength": 32, "maxLength": 512 }, "description": "Required by policy for every command except lock." }
    },
    "responses": {
      "Problem": { "description": "RFC 9457 problem details", "headers": { "X-Correlation-ID": { "schema": { "type": "string" } }, "Retry-After": { "schema": { "type": "integer" } } }, "content": { "application/problem+json": { "schema": { "$ref": "#/components/schemas/Problem" } } } },
...
    },
    "schemas": {
      "Uuid": { "type": "string", "format": "uuid" },
      "Instant": { "type": "string", "format": "date-time" },
...
      "Problem": { "type": "object", "additionalProperties": true, "required": ["type", "title", "status", "code", "correlationId"], "properties": { "type": { "type": "string", "format": "uri-reference" }, "title": { "type": "string" }, "status": { "type": "integer", "minimum": 400, "maximum": 599 }, "detail": { "type": "string" }, "instance": { "type": "string", "format": "uri-reference" }, "code": { "type": "string", "pattern": "^[a-z0-9_]+$" }, "correlationId": { "type": "string" }, "errors": { "type": "array", "items": { "type": "object", "required": ["pointer", "code"], "properties": { "pointer": { "type": "string" }, "code": { "type": "string" } } } } } },
...
      "EmailRequest": { "type": "object", "additionalProperties": false, "required": ["email"], "properties": { "email": { "type": "string", "format": "email", "maxLength": 320 } } },
      "TokenRequest": { "type": "object", "additionalProperties": false, "required": ["token"], "properties": { "token": { "type": "string", "minLength": 32, "maxLength": 512, "format": "password" } } },
      "PasswordResetRequest": { "type": "object", "additionalProperties": false, "required": ["token", "newPassword"], "properties": { "token": { "type": "string", "minLength": 32, "maxLength": 512, "format": "password" }, "newPassword": { "type": "string", "minLength": 12, "maxLength": 1024, "format": "password" } } },
      "PasswordChangeRequest": { "type": "object", "additionalProperties": false, "required": ["currentPassword", "newPassword"], "properties": { "currentPassword": { "type": "string", "minLength": 1, "maxLength": 1024, "format": "password" }, "newPassword": { "type": "string", "minLength": 12, "maxLength": 1024, "format": "password" } } },
      "WebAuthnOptions": { "type": "object", "description": "JSON-serialized PublicKeyCredential creation or request options plus a server challenge ID.", "additionalProperties": true, "required": ["challengeId", "publicKey"], "properties": { "challengeId": { "$ref": "#/components/schemas/Uuid" }, "publicKey": { "type": "object" } } },
...
      "StepUpIntent": { "type": "object", "additionalProperties": false, "required": ["intent", "idempotencyKey"], "properties": { "intent": { "$ref": "#/components/schemas/StepUpIntentType" }, "deviceId": { "oneOf": [{ "$ref": "#/components/schemas/Uuid" }, { "type": "null" }] }, "commandType": { "oneOf": [{ "$ref": "#/components/schemas/CommandType" }, { "type": "null" }] }, "idempotencyKey": { "$ref": "#/components/schemas/Uuid" } } },
      "StepUpIntentType": { "type": "string", "enum": ["command", "account_delete", "data_export", "device_revoke", "security_change"] },
      "StepUpOptions": { "type": "object", "additionalProperties": false, "required": ["intentId", "expiresAt", "methods"], "properties": { "intentId": { "$ref": "#/components/schemas/Uuid" }, "expiresAt": { "$ref": "#/components/schemas/Instant" }, "methods": { "type": "array", "minItems": 1, "uniqueItems": true, "items": { "type": "string", "enum": ["passkey", "password"] } }, "passkeyOptions": { "oneOf": [{ "$ref": "#/components/schemas/WebAuthnOptions" }, { "type": "null" }] } } },
      "StepUpCompletion": { "type": "object", "additionalProperties": false, "required": ["intentId", "method", "proof"], "properties": { "intentId": { "$ref": "#/components/schemas/Uuid" }, "method": { "type": "string", "enum": ["passkey", "password"] }, "proof": { "type": "object", "additionalProperties": true } } },
      "StepUpGrant": { "type": "object", "additionalProperties": false, "required": ["grant", "expiresAt"], "properties": { "grant": { "type": "string", "format": "password" }, "expiresAt": { "$ref": "#/components/schemas/Instant" } } },
      "Profile": { "type": "object", "additionalProperties": false, "required": ["id", "username", "email", "emailVerified", "displayName", "timezone", "marketingOptIn", "createdAt", "version"], "properties": { "id": { "$ref": "#/components/schemas/Uuid" }, "username": { "type": "string" }, "email": { "type": "string", "format": "email" }, "emailVerified": { "type": "boolean" }, "displayName": { "type": "string" }, "dateOfBirth": { "type": ["string", "null"], "format": "date" }, "marketingOptIn": { "type": "boolean" }, "timezone": { "type": "string", "example": "Europe/London" }, "timezoneAssumed": { "type": "boolean" }, "createdAt": { "$ref": "#/components/schemas/Instant" }, "version": { "type": "integer", "minimum": 1 } } },
      "ProfileUpdate": { "type": "object", "additionalProperties": false, "minProperties": 1, "properties": { "displayName": { "type": "string", "minLength": 1, "maxLength": 100 }, "email": { "type": "string", "format": "email", "maxLength": 320 }, "dateOfBirth": { "type": ["string", "null"], "format": "date" }, "marketingOptIn": { "type": "boolean" }, "timezone": { "type": "string", "minLength": 1, "maxLength": 100 }, "expectedVersion": { "type": "integer", "minimum": 1 } } },
      "ExportJob": { "type": "object", "additionalProperties": false, "required": ["id", "status", "createdAt", "expiresAt"], "properties": { "id": { "$ref": "#/components/schemas/Uuid" }, "status": { "type": "string", "enum": ["queued", "processing", "ready", "failed", "expired"] }, "createdAt": { "$ref": "#/components/schemas/Instant" }, "expiresAt": { "$ref": "#/components/schemas/Instant" }, "downloadUrl": { "type": ["string", "null"], "format": "uri-reference" } } },
      "Session": { "type": "object", "additionalProperties": false, "required": ["id", "platform", "clientName", "createdAt", "lastUsedAt", "expiresAt", "current"], "properties": { "id": { "$ref": "#/components/schemas/Uuid" }, "platform": { "$ref": "#/components/schemas/Platform" }, "clientName": { "type": "string" }, "createdAt": { "$ref": "#/components/schemas/Instant" }, "lastUsedAt": { "$ref": "#/components/schemas/Instant" }, "expiresAt": { "$ref": "#/components/schemas/Instant" }, "current": { "type": "boolean" } } },
      "SessionPage": { "allOf": [{ "$ref": "#/components/schemas/Page" }, { "type": "object", "required": ["items"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/Session" } } } }] },
      "Platform": { "type": "string", "enum": ["windows", "android", "ios", "macos", "linux", "web", "unknown"] },
...
      "CommandType": { "type": "string", "enum": ["lock", "sleep", "hibernate", "sign_out", "restart", "shutdown"] },
      "CommandStatus": { "type": "string", "enum": ["queued", "claimed", "accepted", "succeeded", "failed", "expired", "cancelled"] },
      "CommandFailureCode": { "type": "string", "enum": ["no_interactive_session", "unsupported", "permission_denied", "expired", "local_replay", "execution_failed"] },
      "CommandCreate": { "type": "object", "additionalProperties": false, "required": ["type"], "properties": { "type": { "$ref": "#/components/schemas/CommandType" }, "expiresInSeconds": { "type": "integer", "minimum": 1, "maximum": 300, "default": 120 }, "explicitlyConfirmed": { "type": "boolean", "default": false } } },
      "Command": { "type": "object", "additionalProperties": false, "required": ["id", "deviceId", "type", "status", "issuedAt", "expiresAt", "version"], "properties": { "id": { "$ref": "#/components/schemas/Uuid" }, "deviceId": { "$ref": "#/components/schemas/Uuid" }, "type": { "$ref": "#/components/schemas/CommandType" }, "status": { "$ref": "#/components/schemas/CommandStatus" }, "issuedAt": { "$ref": "#/components/schemas/Instant" }, "expiresAt": { "$ref": "#/components/schemas/Instant" }, "claimedUntil": { "oneOf": [{ "$ref": "#/components/schemas/Instant" }, { "type": "null" }] }, "acceptedAt": { "oneOf": [{ "$ref": "#/components/schemas/Instant" }, { "type": "null" }] }, "finishedAt": { "oneOf": [{ "$ref": "#/components/schemas/Instant" }, { "type": "null" }] }, "failureCode": { "oneOf": [{ "$ref": "#/components/schemas/CommandFailureCode" }, { "type": "null" }] }, "version": { "type": "integer", "minimum": 1 } } },
      "CommandClaim": { "type": "object", "additionalProperties": false, "required": ["agentInstanceId"], "properties": { "agentInstanceId": { "$ref": "#/components/schemas/Uuid" } } },
      "CommandAcknowledgement": { "type": "object", "additionalProperties": false, "required": ["state", "agentInstanceId", "localReplayKey"], "properties": { "state": { "type": "string", "enum": ["accepted", "succeeded", "failed"] }, "failureCode": { "oneOf": [{ "$ref": "#/components/schemas/CommandFailureCode" }, { "type": "null" }] }, "agentInstanceId": { "$ref": "#/components/schemas/Uuid" }, "localReplayKey": { "$ref": "#/components/schemas/Uuid" } } },
      "CommandPage": { "allOf": [{ "$ref": "#/components/schemas/Page" }, { "type": "object", "required": ["items"], "properties": { "items": { "type": "array", "items": { "$ref": "#/components/schemas/Command" } } } }] },
...
      "Page": { "type": "object", "additionalProperties": false, "required": ["items", "nextCursor"], "properties": { "items": { "type": "array", "items": true }, "nextCursor": { "type": ["string", "null"] } } }
    }
  },
  "security": [{ "bearerAuth": [] }]
}